Implement the FCS_RCS_STATES partially.
Right now when doing FCS_RCS_STATES, the fcs_state_t * keys are kept
separately from the extra_info_t. The fcs_state_t keys are associated to them
using a map. FCS_RCS_STATES only works with COMPACT_STATES, and then it's
OK.

What now reamains to be done is to prune the fcs_state_t * map and convert
it to an LRU cache.

All tests are successful - with ./Tatzer --fcs. Now we should see that
everything compiles and runs without --fcs.



git-svn-id: file:///home/shlomif/Backup/svn-dumps/google-code/svnsync-repos/fc-solve/branches/rcs-states-storage@3218 e7e8a897-7ba4-4ee7-b36f-f4c66519b19a